@@ -43,7 +43,7 @@ func ScheduleAuthors() {
 	}
 
 	var authors []uint
-	err = db.Model(&widget.Author{}).Where("updated_at < ?", time.Now().Add(-1*time.Hour)).Select("id").Order("updated_at ASC").Limit(500).Find(&authors).Error
+	err = db.Model(&widget.Author{}).Where("updated_at < ?", time.Now().Add(-1*time.Hour)).Select("member_id").Order("updated_at ASC").Limit(500).Find(&authors).Error
 	if err != nil {
 		log.Printf("Failed to pull authors to sync: %s", err)
 		return
@@ -92,7 +92,7 @@ func (consumer *SyncAuthorConsumer) Consume(id uint, ctx context.Context) *widge
 		project := &widget.Project{}
 		//we check for a 403 because the project is "abandoned" and this breaks on the new API
 		//we will assume the list we have is still okay for them
-		err = db.Where("curse_id = ? AND status IN (200, 403)", v.Id).First(project).Error
+		err = db.Where("curse_id = ?", v.Id).First(project).Error
 		if err != nil && err != gorm.ErrRecordNotFound {
 			panic(err)
 		}
